requirements.txt
streamlit
python-dotenv
pyyaml>=6.0.1
reportlab
